djangoapps/learnmore_plus/requirements/prod.txt
-r base.txt

# Production dependencies
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
django-storages>=1.14.2
sentry-sdk>=1.39.1
django-redis>=5.4.0
django-celery-beat>=2.5.0
celery>=5.3.6
redis>=5.0.1
django-cors-headers>=4.3.1 